# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    This is the API documentation for WASAText.
  version: 2.1.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
...
  /groups/{group_id}/name:
    put:
      tags:
        - Groups
      operationId: setGroupName
      summary: Sets the group name
      description: Updates the group's name
      parameters:
      - name: group_id
        in: path
        required: true
        description: ID of the group
        schema:
          $ref: "#/components/schemas/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload containing the new name for the group.
              required: [ name ]
              properties:
                name:
                  $ref: "#/components/schemas/Name"
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group"}
              example:
                id: 1234
                name: "Chill group"
                members:
                  - "alice"
                  - "bob"
        '404':
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/photo:
    put:
      tags:
        - Groups
      operationId: setGroupPhoto
      summary: Sets the group's photo
      description: |
        Updates the group's photo.

      parameters:
        - name: group_id
          in: path
          required: true
          description: ID of the group
          schema:
            $ref: "#/components/schemas/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload containing the new photo of the group.
              required: [ photo ]
              properties:
                photo:
                  $ref: "#/components/schemas/Pic"
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group"}
              example:
                id: 12
                name: "Chill group"
                members:
                  - "alice"
                  - "bob"
                group_photo: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
        '404':
          $ref: "#/components/responses/NotFound"
  /groups/{group_id}/members:
    parameters:
        - name: group_id
          in: path
          required: true
          description: ID of the group
          schema:
            $ref: "#/components/schemas/Id"
    put:
      tags:
        - Groups
      operationId: addToGroup
      summary: Adds one or more users to the group
      description: |
        Adds users to the specified group. Only group members or admins can perform this action.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload to add memebers to a group.
              required: [ member_ids ]
              properties:
                member_ids:
                  type: array
                  description: Array of user IDs to add to the group
                  minItems: 1
                  maxItems: 100
                  items:
                    $ref: "#/components/schemas/Username"
      responses:
        '200':
          description: Members added successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group"}
              example:
                id: 12
                name: "Chill group"
                members:
                  - "alice"
                  - "bob"
                  - "prue"
                  - "phoebe"
                group_photo: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
        '400':
          $ref: "#/components/responses/BadRequest"
        '404':
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Groups
      operationId: leaveGroup
      summary: Removes the current user from the group
      description: |
        The current user leaves the specified group.
      responses:
        '204':
          description: Successfully left the group
        '404':
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id:
      description:  Unique identifier for users, conversations, groups and messages
      type: integer
      format: int64
      minimum: 1
      example: 17
    Username:
      description: Name of a user
      type: string
      pattern: '^[a-zA-Z0-9_.-]{3,13}$'
      minLength: 3
      maxLength: 25
      example: "alice123"
    Name:
      description: Name of a group
      type: string
      pattern: '^[a-zA-Z0-9_ ]+$'
      minLength: 3
      maxLength: 25
      example: "Study group"
...
    Pic:
      description: Base64-encoded image data, it can be JPEG, PNG, GIF, WebP
      type: string
      format: byte
      pattern: '^[A-Za-z0-9+/]+={0,2}$'
      minLength: 100
      maxLength: 500000
      example: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
...
    Group:
      type: object
      description: Group schema.
      required:
        - id
        - name
        - members
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          $ref: "#/components/schemas/Name"
        members:
          type: array
          description: List of user IDs that are members of the group.
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Username"
        group_photo:
          $ref: "#/components/schemas/Pic"
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g., missing or invalid fields).
...
    NotFound:
      description: The requested resource was not found.
...